Adding unit test for telemetry and rearranged some file names and position
