+++ CHANGELOG.md
@@ -1,19 +1,18 @@
-# CHANGELOG (v0.1.4)
+# CHANGELOG (v0.1.5)
 
-## 0.1.4 (2025-02-11)
+## 0.1.5 ()
 
-### Backwards incompatible changes for 0.1.3
+### Backwards incompatible changes for 0.1.4
  * None
 
 ### Bug fixes
  * None
 
 ### Enhancements
- * [[`PR-6`](https://github.com/thiagoesteves/observer_web/pull/6)] Adjusted Observer Web Font
- * [[`PR-7`](https://github.com/thiagoesteves/observer_web/pull/7)] Adding Live Metrics in Oberver Web for capturing the VM statistics using OTP distribution
  * [[`PR-8`](https://github.com/thiagoesteves/observer_web/pull/8)] Adding Metrics using telemetry
 
 # Previous Releases
+ * [0.1.4 🚀 (2025-02-11)](https://github.com/thiagoesteves/observer_web/blob/v0.1.4/CHANGELOG.md)
  * [0.1.3 🚀 (2025-02-08)](https://github.com/thiagoesteves/observer_web/blob/v0.1.3/CHANGELOG.md)
  * [0.1.2 🚀 (2025-02-08)](https://github.com/thiagoesteves/observer_web/blob/v0.1.2/CHANGELOG.md)
  * [0.1.0 🚀 (2025-01-06)](https://github.com/thiagoesteves/observer_web/blob/v0.1.0/CHANGELOG.md)
