+++ lib/web/telemetry.ex
@@ -1,6 +1,7 @@
 defmodule Observer.Web.Telemetry do
   use Supervisor
   import Telemetry.Metrics
+  import ObserverWeb.Macros
 
   def start_link(arg) do
     Supervisor.start_link(__MODULE__, arg, name: __MODULE__)
@@ -8,15 +9,25 @@ defmodule Observer.Web.Telemetry do
 
   @impl true
   def init(_arg) do
-    children = [
-      # Telemetry poller will execute the given period measurements
-      # every 10_000ms. Learn more here: https://hexdocs.pm/telemetry_metrics
-      {:telemetry_poller, measurements: periodic_measurements(), period: 5_000},
-      # Add reporters as children of your supervision tree.
-      {Observer.Web.TelemetryLocal, metrics: metrics()}
-    ]
+    # Telemetry poller will execute the given period measurements
+    # every 5_000ms. Learn more here: https://hexdocs.pm/telemetry_metrics
+    children =
+      add_telemetry_poller() ++
+        [
+          # Add reporters as children of your supervision tree.
+          {ObserverWeb.Telemetry.Consumer, metrics: metrics()}
+        ]
 
     Supervisor.init(children, strategy: :one_for_one)
+  end
+
+  if_not_test do
+    defp add_telemetry_poller,
+      do: [
+        {:telemetry_poller, measurements: periodic_measurements(), period: 5_000}
+      ]
+  else
+    defp add_telemetry_poller, do: []
   end
 
   def metrics do
@@ -61,11 +72,13 @@ defmodule Observer.Web.Telemetry do
     ]
   end
 
-  defp periodic_measurements do
-    [
-      # A module, function and arguments to be invoked periodically.
-      # This function must call :telemetry.execute/3 and a metric must be added above.
-      {ObserverWeb.Telemetry.Producer.PhxLvSocket, :process_phoenix_liveview_sockets, []}
-    ]
+  if_not_test do
+    defp periodic_measurements do
+      [
+        # A module, function and arguments to be invoked periodically.
+        # This function must call :telemetry.execute/3 and a metric must be added above.
+        {ObserverWeb.Telemetry.Producer.PhxLvSocket, :process_phoenix_liveview_sockets, []}
+      ]
+    end
   end
 end
