+++ test/observer_web/telemetry_storage_test.exs
@@ -1,10 +1,10 @@
-defmodule ObserverWeb.TelemetryTest do
+defmodule ObserverWeb.TelemetryStorageTest do
   use ExUnit.Case, async: false
 
   import Mock
   import Mox
 
-  alias ObserverWeb.Telemetry.Consumer, as: Consumer
+  alias ObserverWeb.Telemetry.Storage
   alias ObserverWeb.TelemetryFixtures
 
   setup [
@@ -14,7 +14,7 @@ defmodule ObserverWeb.TelemetryTest do
   ]
 
   test "[un]subscribe_for_new_keys/0", %{node: node} do
-    Consumer.subscribe_for_new_keys()
+    Storage.subscribe_for_new_keys()
 
     ObserverWeb.RpcMock
     |> stub(:call, fn ^node, :ets, :lookup, [:observer_web_metrics, "metric-keys"], :infinity ->
@@ -23,13 +23,13 @@ defmodule ObserverWeb.TelemetryTest do
 
     node
     |> TelemetryFixtures.build_reporter_vm_memory_total()
-    |> Consumer.push_data()
+    |> Storage.push_data()
 
     assert_receive {:metrics_new_keys, ^node, ["vm.memory.total"]}, 1_000
   end
 
   test "[un]subscribe_for_new_data/0", %{node: node} do
-    Consumer.subscribe_for_new_data(node, "vm.memory.total")
+    Storage.subscribe_for_new_data(node, "vm.memory.total")
 
     ObserverWeb.RpcMock
     |> stub(:call, fn ^node, :ets, :lookup, [:observer_web_metrics, "metric-keys"], :infinity ->
@@ -38,18 +38,18 @@ defmodule ObserverWeb.TelemetryTest do
 
     node
     |> TelemetryFixtures.build_reporter_vm_memory_total()
-    |> Consumer.push_data()
+    |> Storage.push_data()
 
     assert_receive {:metrics_new_data, ^node, "vm.memory.total",
                     %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: _, measurements: _}},
                    1_000
 
     # Validate by inspection
-    Consumer.unsubscribe_for_new_data(node, "vm.memory.total")
+    Storage.unsubscribe_for_new_data(node, "vm.memory.total")
   end
 
   test "get_keys_by_node/1 valid node", %{node: node} do
-    Consumer.subscribe_for_new_data(node, "vm.memory.total")
+    Storage.subscribe_for_new_data(node, "vm.memory.total")
     test_pid = self()
 
     ObserverWeb.RpcMock
@@ -65,23 +65,23 @@ defmodule ObserverWeb.TelemetryTest do
 
     node
     |> TelemetryFixtures.build_reporter_vm_memory_total()
-    |> Consumer.push_data()
+    |> Storage.push_data()
 
     assert_receive {:metrics_new_data, ^node, "vm.memory.total",
                     %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: _, measurements: _}},
                    1_000
 
-    assert ["vm.memory.total"] == Consumer.get_keys_by_node(node)
+    assert ["vm.memory.total"] == Storage.get_keys_by_node(node)
   end
 
   test "node/1 invalid node" do
-    assert [] == Consumer.get_keys_by_node(nil)
+    assert [] == Storage.get_keys_by_node(nil)
   end
 
   test "list_data_by_node_key/3", %{node: node} do
     key_name = "test.phoenix"
 
-    Consumer.subscribe_for_new_data(node, key_name)
+    Storage.subscribe_for_new_data(node, key_name)
 
     ObserverWeb.RpcMock
     |> stub(
@@ -104,7 +104,7 @@ defmodule ObserverWeb.TelemetryTest do
       end
     )
 
-    Enum.each(1..5, &Consumer.push_data(build_metric(node, key_name, &1)))
+    Enum.each(1..5, &Storage.push_data(build_metric(node, key_name, &1)))
 
     assert_receive {:metrics_new_data, ^node, ^key_name, %{timestamp: _, unit: _, value: 5}},
                    1_000
@@ -115,7 +115,7 @@ defmodule ObserverWeb.TelemetryTest do
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 3, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 4, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 5, tags: _}
-           ] = Consumer.list_data_by_node_key(node |> to_string(), key_name, order: :asc)
+           ] = Storage.list_data_by_node_key(node |> to_string(), key_name, order: :asc)
 
     assert [
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 5, tags: _},
@@ -123,7 +123,7 @@ defmodule ObserverWeb.TelemetryTest do
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 3, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 2, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 1, tags: _}
-           ] = Consumer.list_data_by_node_key(node |> to_string(), key_name, order: :desc)
+           ] = Storage.list_data_by_node_key(node |> to_string(), key_name, order: :desc)
 
     assert [
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 1, tags: _},
@@ -131,7 +131,7 @@ defmodule ObserverWeb.TelemetryTest do
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 3, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 4, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 5, tags: _}
-           ] = Consumer.list_data_by_node_key(node |> to_string(), key_name)
+           ] = Storage.list_data_by_node_key(node |> to_string(), key_name)
   end
 
   test "Pruning expiring entries", %{node: node, pid: pid} do
@@ -139,7 +139,7 @@ defmodule ObserverWeb.TelemetryTest do
 
     now = System.os_time(:millisecond)
 
-    Consumer.subscribe_for_new_data(node, key_name)
+    Storage.subscribe_for_new_data(node, key_name)
 
     ObserverWeb.RpcMock
     |> stub(
@@ -163,7 +163,7 @@ defmodule ObserverWeb.TelemetryTest do
     )
 
     with_mock System, os_time: fn _ -> now - 120_000 end do
-      Enum.each(1..5, &Consumer.push_data(build_metric(node, key_name, &1)))
+      Enum.each(1..5, &Storage.push_data(build_metric(node, key_name, &1)))
 
       assert_receive {:metrics_new_data, ^node, ^key_name, %{timestamp: _, unit: _, value: 5}},
                      1_000
@@ -175,12 +175,12 @@ defmodule ObserverWeb.TelemetryTest do
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 3, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 4, tags: _},
              %ObserverWeb.Telemetry.Data{timestamp: _, unit: _, value: 5, tags: _}
-           ] = Consumer.list_data_by_node_key(node |> to_string(), key_name, order: :asc)
+           ] = Storage.list_data_by_node_key(node |> to_string(), key_name, order: :asc)
 
     send(pid, :prune_expired_entries)
     :timer.sleep(100)
 
-    assert [] = Consumer.list_data_by_node_key(node |> to_string(), key_name, order: :asc)
+    assert [] = Storage.list_data_by_node_key(node |> to_string(), key_name, order: :asc)
   end
 
   defp build_metric(node, name, value) do
@@ -202,7 +202,7 @@ defmodule ObserverWeb.TelemetryTest do
 
   defp create_consumer(context) do
     node = Node.self()
-    {:ok, pid} = Consumer.start_link([])
+    {:ok, pid} = Storage.start_link([])
 
     context
     |> Map.put(:node, node)
